superficie -> longueur bigDecimal -> int

+++ src/main/java/fr/formation/model/Habitation.java
@@ -38,7 +38,7 @@ public class Habitation extends Batiment {
 	public Habitation() {
 	}
 
-	public Habitation(BigDecimal superficie, String nom, BigDecimal prix, BigDecimal coutEntretienBase, int nbPlace,
+	public Habitation(int superficie, String nom, BigDecimal prix, BigDecimal coutEntretienBase, int nbPlace,
 			Coordonnees coordonnees, Biome biome, Proprietaire proprietaire, BigDecimal loyer) {
 		super(superficie, nom, prix, coutEntretienBase, nbPlace, coordonnees, biome, proprietaire);
 		this.loyer = loyer;
